[WIP] STREAMPIPS-583 working on download service
