+++ ui/src/app/configuration/datalake-configuration/datalake-configuration.component.ts
@@ -161,7 +161,9 @@ export class DatalakeConfigurationComponent implements OnInit {
       title: 'Download data',
       width: '50vw',
       data: {
-        'measureName': measurementName,
+        'dataDownloadDialogModel': {
+          'measureName': measurementName,
+        }
       }
     });
   }
