+++ ui/src/app/core-ui/data-download-dialog/components/select-data/select-data-missing-values/select-data-missing-values.component.scss
@@ -15,19 +15,5 @@
  *  limitations under the License.
  *
  */
-.sp-radio-group {
-  display: flex;
-  flex-direction: column;
-  margin: 15px 0;
-}
 
-.sp-radio-button {
-  margin: 5px;
-  width: 100%;
-}
-
-.form-field-date {
-  top: -10px;
-  width: 300px;
-}
