+++ ui/src/app/core-ui/data-download-dialog/components/select-data/select-data-range/select-data-range.component.scss
@@ -1,4 +1,4 @@
-/*
+/*!
  *  Licensed to the Apache Software Foundation (ASF) under one or more
  *  contributor license agreements.  See the NOTICE file distributed with
  *  this work for additional information regarding copyright ownership.
@@ -16,21 +16,3 @@
  *
  */
 
-import { Component, Input, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'sp-select-data-missing-values',
-  templateUrl: './select-data-missing-values.component.html',
-  styleUrls: ['./select-data-missing-values.component.scss']
-})
-export class SelectDataMissingValuesComponent implements OnInit {
-
-  @Input() missingValueBehaviour: 'ignore' | 'default' | 'empty';
-
-  constructor() {
-  }
-
-  ngOnInit(): void {
-  }
-
-}
