+++ ui/src/app/core-ui/data-download-dialog/components/select-data/select-data.component.html
@@ -16,4 +16,13 @@
   ~
   -->
 
-<p>select-format works!</p>
+<div class="mt-10">
+  <sp-select-data-range
+      [dataExplorerDataConfig]="dataExplorerDataConfig"
+      [dataExportConfig]="dataExportConfig">
+  </sp-select-data-range>
+
+  <sp-select-data-missing-values
+      [dataExportConfig]="dataExportConfig">
+  </sp-select-data-missing-values>
+</div>
