+++ ui/src/app/core-ui/data-download-dialog/components/select-data/select-data.component.scss
@@ -16,7 +16,7 @@
  *
  */
 
-@import 'src/scss/sp/sp-dialog';
+@import '../../../../../scss/sp/sp-dialog';
 
 .sp-radio-group {
   display: flex;
