+++ ui/src/app/core-ui/data-download-dialog/components/select-format/select-format.component.html
@@ -16,14 +16,22 @@
   ~
   -->
 
-<div class="mt-10">
-  <sp-select-data-range
-      [date]="date"
-      [dataConfig]="dataConfig"
-      [dataRangeConfiguration]="exportDataConfiguration.dataRangeConfiguration">
-  </sp-select-data-range>
-
-  <sp-select-data-missing-values
-      [missingValueBehaviour]="exportDataConfiguration.missingValueBehaviour">
-  </sp-select-data-missing-values>
+<ng-template matStepLabel>Select Format</ng-template>
+<div>
+  <h5>Download Format</h5>
+  <mat-radio-group class="sp-radio-group" [(ngModel)]="formatExportConfig.exportFormat">
+    <mat-radio-button value="json" class="sp-radio-button">
+      JSON
+    </mat-radio-button>
+    <mat-radio-button value="csv" class="sp-radio-button">
+      CSV
+    </mat-radio-button>
+  </mat-radio-group>
+</div>
+<div *ngIf="formatExportConfig.exportFormat === 'csv'" class="mt-10">
+  <h5>Delimiter</h5>
+  <mat-radio-group [(ngModel)]="formatExportConfig.delimiter" class="sp-radio-group">
+    <mat-radio-button value="comma" class="sp-radio-button">&nbsp;,</mat-radio-button>
+    <mat-radio-button value="semicolon" class="sp-radio-button">&nbsp;;</mat-radio-button>
+  </mat-radio-group>
 </div>
