+++ ui/src/app/core-ui/data-download-dialog/components/select-format/select-format.component.scss
