--- ui/src/app/core-ui/data-download-dialog/select-data/select-data-range/select-data-range.component.html
@@ -1,86 +1,0 @@
-<!--
-  ~  Licensed to the Apache Software Foundation (ASF) under one or more
-  ~  contributor license agreements.  See the NOTICE file distributed with
-  ~  this work for additional information regarding copyright ownership.
-  ~  The ASF licenses this file to You under the Apache License, Version 2.0
-  ~  (the "License"); you may not use this file except in compliance with
-  ~  the License.  You may obtain a copy of the License at
-  ~
-  ~     http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~  Unless required by applicable law or agreed to in writing, software
-  ~  distributed under the License is distributed on an "AS IS" BASIS,
-  ~  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~  See the License for the specific language governing permissions and
-  ~  limitations under the License.
-  ~
-  -->
-
-<h5>Data Range</h5>
-<mat-radio-group
-    class="sp-radio-group"
-    [ngModel]="dataRangeConfiguration">
-  <mat-radio-button
-      value="visible"
-      class="sp-radio-button"
-      *ngIf="dataConfig">
-    Currently configured query
-  </mat-radio-button>
-  <div fxLayout="column"
-       fxLayoutAlign="start start"
-       class="ml-35 mb-10"
-       *ngIf="dataRangeConfiguration === 'visible' && dataConfig && dataConfig.sourceConfigs.length > 1">
-    <h5>Select source (only one source can be exported in a single file)</h5>
-    <mat-radio-group
-        class="sp-radio-group"
-        fxLayout="column">
-      <!--        [ngModel]="ngModel_1"-->
-      <!--        (ngModelChange)="ngModel_1Change.emit($event)"-->
-      <mat-radio-button
-          class="p-5"
-          [value]="i"
-          *ngFor="let sourceConfig of dataConfig.sourceConfigs; let i = index">
-        {{sourceConfig.measureName}}
-      </mat-radio-button>
-    </mat-radio-group>
-  </div>
-  <mat-radio-button
-      value="all"
-      class="sp-radio-button">
-    All data in database
-  </mat-radio-button>
-  <mat-radio-button
-      value="customInterval"
-      class="sp-radio-button">
-    All data in custom time interval
-  </mat-radio-button>
-</mat-radio-group>
-<div
-    fxLayout="row"
-    fxLayoutAlign="start center"
-    class="ml-35"
-    *ngIf="dataRangeConfiguration === 'customInterval'">
-  <mat-form-field
-      class="form-field-date">
-    <input
-        matInput
-        [owlDateTime]="dt1"
-        [owlDateTimeTrigger]="dt1"
-        [(ngModel)]="dateRange"
-        [selectMode]="'range'"
-        [disabled]="dataRangeConfiguration !== 'customInterval'">
-    <mat-icon
-        matSuffix
-        [owlDateTimeTrigger]="dt1"
-        *ngIf="dataRangeConfiguration === 'customInterval'">
-      event
-    </mat-icon>
-    <mat-icon
-        matSuffix
-        class="event-color"
-        *ngIf="dataRangeConfiguration !== 'customInterval'">
-      event
-    </mat-icon>
-    <owl-date-time #dt1></owl-date-time>
-  </mat-form-field>
-</div>
